pyproject.toml
[tool.poetry]
name = "todam-promptflow"
version = "0.1.0"
description = ""
authors = ["sh1un <a0923183408@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"
langchain-openai = "^0.1.3"
langchain = "^0.1.16"
promptflow = "^1.9.0"
promptflow-tools = "^1.4.0"
langchain-pinecone = "^0.1.0"
python-dotenv = "^1.0.1"
anthropic = {extras = ["bedrock"], version = "^0.25.8"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
